requirements.txt
einops==0.8.1
lpips==0.1.4
matplotlib==3.7.2
numpy==1.24.3
OpenEXR==1.3.9
Pillow==11.1.0
pytorch_lightning==2.0.3
PyYAML==6.0.1
scipy==1.10.1
opencv-python
pathlib